requirements.txt
python3-django
python3-djangorestframework
python3-psycopg2
